fix(backend): widen native_retry_attempt.attempt to integer (cubic)
The smallint column was cast to/from u32 and could wrap a retry chain longer than
i16::MAX into premature exhaustion. Use integer, matching the retry policy's i32
attempt count, so no narrowing occurs on the maybe_enqueue read/write path.

+++ backend/windmill-queue/src/jobs.rs
@@ -1865,7 +1865,7 @@ pub async fn maybe_enqueue_native_script_retry(
     sqlx::query!(
         "INSERT INTO native_retry_attempt (job_id, attempt) VALUES ($1, $2)",
         new_id,
-        (prev_attempts + 1) as i16,
+        (prev_attempts + 1) as i32,
     )
     .execute(&mut *tx)
     .await?;
